Rendered the README.Rmd

+++ tests/testthat/test-app.R
@@ -1,3 +1,3 @@
 test_that("random_app works", {
-  expect_is(random_num(), "numeric")
+  expect_is(random_num(), "integer")
 })
